Move current REST API into a plugin.
And mark it as "legacy". This makes Core Restlet-free, plexus-restlet-bridge free
and life better than before.

Any plugin contributing REST API needs to declare a dependency on this new plugin
for now, but if we introduce some other REST imple, by updating resources in plugins
the dependency will cease too. Ultimately, nexus-legacy-restlet-api-plugin will
remain alone, with no dependant, and is easily removed.

Also, since Core classloader now does not have Restlet, plexus-restlet, and
any other "webtech" loaded (is not a real headless app), the doors are open
to introduce any other remoting we want. But, as a plugin, to keep us
away from mistakes in past!

Also, while security is needed in core, and is in core, the one module
security-rest-api, that is adding security rest apis is now also in this
plugin.

Velocity TCCL muck is handled in 1.23 of restlet bridge.