NEXUS-5354: Fix for concurrency issue
With previous fix, it was possible that 2nd thread trying to repair,
that was being blocked during the repair, would try to also repair an
already repaired index.

Now, indexer will not only blow with IOEx, but will also flag itself as
stopped, and check is added to repair only if indexer is stopped. Hence,
the 2nd thread from above will not do anything.