Fixes needed for proper runtme deps
Due to move of REST API into a plugin, test classpath
left incomplete. Due nexus plugins, the scope is
provided, but as such, all the provided deps (otherwise in core)
are not coming in transitively, hence, runtime classpath
(in this case for UTs/ITs only) are incomplete.

Fix was to add dependency to nexus-web-utils, where needed
classes reside, but also it will pull in needed shiro
deps (most typically shiro-web) that is in core, but the
plugin in question (like rrb or ldap) has no and should
have no knowledge or dependency decleared on them directly,
as they are runtime, indirectly needed dependencies for
security to work.