LVO _is_ and always _was_ a plugin.
But, it was never used as such (JAR was used instead of bundle).
This branch now fixes this, as LVO is from now on used as a real
plugin, and actually this is a must from now on, as REST is
in plugin and LVO exposes REST API. To access restlet1x plugin,
one must to be a plugin too.