Timeline index repair failure is not propagated.
Timeline handles indexer with special care: it will try
to repair it if it's start failed. This change now
will not propagate anymore the exception thrown during
repair (already in exception handler!), but will mark the
indexer as "dead", log the fact, and continue the usual
fluff.

On next reboot, another repair will be tried.