NEXUS-5368: Stale connections in pool
Explicitly controlling the duration for how long
is connection kept alive (it's total lifespan).

Some servers announce their "wish" for how long to
keep connection alive with non standard "Keep-Alive"
response header, and HC4 obeys that header. But some
servers does not do this (like RSO) and are silently
dropping the connection, while HC4 in this case
sets expiration time of the connection to "indefinite"
(forever).

With this change, we take over explicit control over
keep alive duration, and we prevent HC4 to pool connections
forever.