NEXUS-5400: Simplifying.
Original "fix" was to put the uploaded file into "tmp" repo folder having
same directory structure as "final" file would have. This is completely
unnecessary!

We not just store the file in the root of repo temp folder (using
File.createTempFile guarantees uniqueness), as it will be moved
into its finale place in equally same "cheap" way as it would be
in it's intended subfolder.