NEXUS-5400: Modified unrelated IT to be more tolerant for "working files"
This it was eagerly searching for empty directories in all repositories
and it is testing evict tasks, and the fact that emopty folders
are removed from storage after eviction. Sadly, this IT
checks _whole_ structure, including "special" folders like ".meta" and
".nexus" now, as this IT was added way before those folders popped up.

This change now makes this IT to neglect ".nexus/tmp" folder, but
it might be broadened too, since it is interested in empty _artifact_
folders, but not in empty attributes, empty metadata or empty temp
folders.