Fixing the build.
forge parent 27 introduced this change:
https://github.com/sonatype/oss-parents/commit/9261d095a053d43f0bcc827fc88383b4a1513164

But it breaks Nexus build for following reason:
modules nexus-plugin-api and nexus-plugin-test-api are
"grouping" POMs (provided all is needed for plugin dev and
plugin testing) but themselves does not have any content yet.

Adding content was once discussed (moving in plugin-api and plugin-host-api etc)
but did not happen.

So, what happens, is that reactor build will not have main JAR artifact
for these (as they are empty and parent pom instructs JAR plugin to not
create empty JARs), but all the plugins are referencing these modules
as JARs. In case of CLI build using "-U" you will see that the JAR
is checked/fetched from remote. Zion on the other hand, simply
fails, as it is isolated environment, and will not have the JAR
accessible by any means.