NEXUS-5422 use String to determine if two versions are equal
Aether uses Maven version comparison rules to determine if two
versions are equal, and considers 2 and 2.0 to be the same version,
which is not appropriate for index searches.

Changed the code to use aether logic to order versions, but fall
back to string comparison to determine if versions are equal or not.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>