Nexus570IndexArchetypeIT fix, take two.
Last failure:
https://bamboo.zion.sonatype.com/browse/NXO-OSST-118

This time a small change implemented in AbstractNexusIntegrationTest
and added proper packaging to the test project being
deployed as part of preparations for this IT.

Original assumption still stands, asynchronously run
event inspector ordering is not guaranteed, and also
reliability of "wait for ..." IT utilities are questioned
too (NEXUS-5127).

With this change (redone), it should never fail again,
as it will immediately (on POM arrival) be indexed as archetype,
not JAR anymore.