NEXUS-5472: UT for WL maintenance on hosted repositories
And a fix, as it immediately revealed how WL maintenance lacks
proper functionality when deletion happens in hosted repository.
Problem is that when artifact is removed directly, Nexus will
leave (potentially empty) parent folders untouched, and no events
will be emitted either for those (to remove WL entry "/org/sonatype")
for example. To process the delete event, get the matching entry,
and than to decide should the WL entry be removed or not, a storage
scanning might be needed, as "org/sonatype" might cover zillion other
GAVs, not the single one deleted. In this case, currently, NO WL entry
removal will happen.

The case when given path (that is in WL) or it's parent is deleted
works now and is covered with this UT.