NEXUS-5472: Legacy ITs fix
36 legacy ITs are failing now due to WL. Most
of them are setting up nexus and then remote proxy,
and WL causes to auto block proxies in test, causing
IT to fail due to items not found. Others
are counting remote proxy hits, and failing due to
unexpected WL discovery requests.

In general, as all the "legacy" ITs are coming from
pre-WL era, simplest fix was to simply disable WL
for all of them.

New WL ITs will use the new IT infrastructure anyway.