NEXUS-5472: Last bits to fix wrt legacy UTs and OOM
Last night's fix was not enough. So, reworked the "flag"
that was in EventDispatcher, and now it's possible to disable
the feature in consistent way (everything, even REST and UI) will
work but no remote content discovery will be possible.

This flag is not meant for production, only in some rare
case of troubleshooting. It's actual purpose is use
in "legacy" UTs and ITs that are unaware of this
new feature and it's async jobs and requests.