Related to remote-Catch-All change.
Simple change to ensure no "badly protected" resource
get mounted, by simply changing the order of security
handing and actual mount.

As now, in Nexus we do check are resources protected
properly, and throwing exception if not, it was resulting
that resource get mounted but remained unprotected.

Now, we ensure no badly protected resource is actually
attached to Restlet application.