NEXUS-5472: Making sure REST emits keys even if they are empty
Specifically in case of "In progress" for discovery, lastMessage
and lastStrategy were left as nulls, and UI (or ExtJS code) was
not updating corresponding fields as they were not present in the
response message at all.

Now we ensure that those fields will be updated (when you press
Force remote discovery with previously ran discovery), and
status fields will be updated.