Improving ITs
Current ITs seems unstable due to that they are unable to properly
detect when and what async operation is done on server. To improve
that introduced new method that listens for publishing too, as
previous failures were due to central being discovered, but public
group prefix file was still being built, and asserts were failing.