NEXUS-5602: Fixing the test
The code is correct but test was incorrect as it asserted
against one possible outcome of updates (due to concurrency).

The point is that -- depending how bg jobs were
executed -- group was updated one or twice. One thing is
sure, group has to be updated AT LEAST ONCE.

Adjusted assertions accordingly, to not make this test
fail in case when group is updated only once, as that is
also valid outcome.