NEXUS-5618: M1 reposes are polluting logs on boot
The init step of WLManager did not change for Maven2
content class, only for the MavenRepo "facet". This
is not enough, as on boot M1 reposes will
be tried to init wrongfully. That is not
causing any problem except polluting logs
with stack traces.

We had something similar already fixed, but it was
about repositories put out of service...