NEXUS-5623: Hosted WL not updated on parent removal
WL for hosted repositories is not properly maintained
when a parent directory is removed (parent of WL entry).

For example, when your WL has entry like "/org/sonatype"
and you delete folder "/org".

Moreover, the UT testing this was broken too, as it was
testing this but actually missed the point: comments
stated one thing but assertions were asserting the
opposite