NEXUS-5612: File listing should fail in case of IO problem
Currently DefaultFSPeer was hiding the null returned
from File.listFiles. According to javadocs, it returns null
in two cases, a) when the File instance does not denote
a directory, or b) in case of IO error.

As case a) is checked few lines above (if check failed,
we would not get here), the null return value should
be treated as IO Error.

By throwing a generic LocalStorageExcetion, we
make this case to be treated as error.