NEXUS-5602: Using the periodic update thread for initial updates too
On 1st boot of updated instances, a lot of MavenRepository
instances (hosted and proxy) will be piled up in the array
needUpdateRepositories in WLManagerImpl#startup().

Currently, as the thread pool with fixed size of 5 and
blocking queue was deferring update.

With this change, the for-loop doing initial WL build
(walk of the content for hosted and discovery for proxies)
is deferred, and periodic updater thread is used.

This also means, that upon boot, the initial delay of
first run will be applied, and on that moment on, hosted
reposes will have WL, not before that.

Initial change breaks all the UTs and ITs that uses
exposed-for-tests-only method to reach a state when
the test should start.

isUpdateWhitelistJobRunning() method used for this
is now updated, and will flag (and make tests wait)
as long as boot sequence is done (newly added) AND
there are no bg updates running (as before).

Also, EventDispatcher simplified, the feature active
flag is not passed anymore, instead, it is not
registered to event bus when feature is not
active, resulting in same behavior (before, when
feature was inactive EventDispatcher got events
but simply did not invoke WL Manager).