Fixed wrong RoutingStatusResourceTest initialization sequence
The test was waiting for routing background updates to finish
way too early (before nexus repo manager was started) and was
extremely sensitive to changes auto-routing implementation.

Changed the test to do waiting later during nexus instance
startup. This is still a workaround and the true fix should
wait for a particular event, "repo auto-routing publish state
change" in this case, not for "no backgroup update jobs are
running" as it does now.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>