NEXUS-5612: Testing interaction with WalkerProcessor and FSPeer
Especially the case when FSPeer#listItems method throws
LocalStorageException.

The Walk must be stopped, after this, no processor
should have onCollectionExit invoked.

This UT actually simulates the case when due to
too many open files, FSPeer made Walker think
that collection is empty, while it was not.

DefaultFSPeer is already fixed in prev commit,
this UT only ensures that Walker, Context and
Processors behave correctly in this case.