NXCM-5153: Core changes needed
This abstract REST resource is used by Staging
Upload resource too. These are the changes needed
to implement what we need for staging.

Changes:
* reverse engineered and doc'd how upload actually works
* variable and their scope cleanup
* minor logging cleanup

The key change is introduction of new method
uploadGavParametersAvailable that is in this same
class implemented as nop. This method is called by
upload method only once, at the moment when
upload GAVP coordinates are available.