NXCM-5188: NPE fix
In the issue attached log, it's clearly visible that validation
did reject the item (policy=Strict but no hashes were present).
Later, an NPE is thrown, as the cleanup method is not protected
from this case.

In all other case it would work as contentValid would be
TRUE for all of those. This is the only branch in logic
where it is FALSE, and the remoteHash is null too.

In other words, cleanup is needed when content is
invalid, but hashes arrived into cache.