NXCM-5194: G level metadata merge duplicates entries
Reason was as entry equality was checked by comparing all
three properties of the Plugin: prefix, artifactId and name.

This commit "improves" this, by performing comparison
on prefix and artifactId only. This means that
duplication will not happen anymore on new
deploys, and will be corrected if maven metadata
rebuild task is run against repository. Still,
the one case "changing prefix" mentioned
in related issue is NOT fixed by this change.

Related:
https://issues.sonatype.org/browse/NEXUS-4448