Fix for intermittent NPE in routing tests.
Problem was that UT/ITs had no more means to detect
when "boot is done", and sometimes they kicked in
too early, with expectations (or without proper
checks) causing NPEs with carelessly loading
up prefix files without checking for their
existence (as "they should exist").

Reason was that boot process of routing
did not finish yet, as in 1st pass all
the prefix files are created for hosted
repositories for example.