NXCM-4884: Stripped off authz completely.
As for one, it never worked. As for two, this was
the reason (swallowed JSONException) that prevented
proper login to happen.

authz with newly introduced POST was working, but
building the role list failed, as Kenai does
not answers what authz code in this realm was
expecting (a JSON string). for now, simply
the default role (set in kenai XML config)
is returned.