Forcing the binding type
These two descendants of ManagedPlexusResource are the only
non-plexus components (the other ones are all Plexus comps).
This resulted due to Guice "looser" bindings to have them end
up in collection of PlexusResources too as the two ifaces
are type compatible.

Explicitly setting the type on these two for now.