Update downloadRepositoryIndex to conform to new indexer semantics
The maven indexer semantics were changed to fail if an incremental only
update was requested, but not possible. Previously, it would move on to
a full index update if an incremental failed. downloadRepositoryIndex
needed to be updated to account for this. To keep the behavior the same,
downloadRepositoryIndex will handle a failed incremental only update,
and try a full index update.