Fix failing integration tests
Due to the introduction of DefaultSecurityPasswordService,
security-configuration.xml was now getting loaded prior to the test
security-configuration.xml file being copied into the proper location.
This caused the defaults to get loaded, and subsequently overwrite the
test security-configuration.xml with the defaults. The tests were then
being run with the defaults, which caused the failures.

Force a load of security-configuration.xml after the test file is in
place