NEXUS-5710: Status and Content CSS files were not loaded
The URL generated by template (content and status template)
were always wrong, it contained "//" between nexusRoot
and rest of the path info.

But the changes how the resource is served up now
made the "//" appear as error, Restlet refuses
to serve it up (was 404).

The two places of use of getContextRoot(req) where
templates are evaluated are fixed (third place related
to openSearch already had a fix for this), as all the
other users of above mentioned method EXPECT that
returned reference contains path info ending with "/",
unlike the cases related to Velocity templates.