Apply the common pattern for model exclusion.
Similarly as in Pro, here the indexer model
depends on restlet1x model, and it lands in
the plugin bundle (which is wrong).

Added exclusion to model dependency as the
"real" path to bring it in is via plugin
dependency (indexer depends on restlet1x plugin)
and hence, the scope is correct this way too
(provided vs compile).