NEXUS-5739: Remove HC3
Removed HC3, RRS imple using it and related
tests that were testing HC3 RRS or the customized
CustomMultiThreadedHttpConnectionManager for
connection leaks and non-proxy-host support.

Still, legacy ITs nexus-test-harness-launcher
still has it as dependency, as many
legacy "clients" and "message util" classes
has direct dependency on it.

At the end, HC3 is out of the bundle, but
is still used by legacy ITs only.