NEXUS-5744: fix for redirect strategy
There was initially a "prevent redirection to index pages"
check that proved to be overly broad. It was meant to prevent
redirections like:

http://host/path -> http://host/path/

as "path" is actually a collection (directory in HTTP lingo)
on server side, and this is usual redirections that HTTP
servers do when asked for that 1st URL.

But, this check was too broad, as it prevented redirections
like

http://hostA/path/ -> http://hostB/path/

for example (same for HTTP -> HTTPS or even port change),
as it was checking only that 2nd URL does not end with slash.