Fixing ITs.
Nexus4301WarnErrorLogsLoadTestIT created a LOT of entries,
that might not all got picked up due to default of 1000 log lines
to be processed. Using enhanced feed util to instruct the feed
to process all available logfile, to make sure all the
generated entries are picked up.

Nexus4427WarnErrorLogsToFeedsIT asserted with "startsWith",
but as currently "raw" log lines are picked up, it was changed
to "contains" assertion.

But, the latter IT still fails, as it explicitly checks
for Jetty EofException and TimelineException being
filtered out from feed, something that does not happen
anymore.

IMO, is unnecessary, as these filterngs were just a
workaround for logging problem in Restlet, fixed in
patched Restlet V8 we use currently. As for TimelineEx,
it was also fixed, was due to misuse of Lucene in Timeline