Making IT more reliable.
By inserting a new waitForAllTasksToStop. The
runTask should wait for the spawned bg task,
but it seems unreliable, and is some cases
the call does not block.

By having the 2nd call, we actually increase the
probability that server side task actually get
running, hence, be detected (and block IT main thread)
and actually be awaited to finish it's job.