NXCM-5396: Realms that have no UserManager pollute logs
This change lessens the log noise, while the information
about "not found user" is still logged.

On instances where some Realm is used that does
not have UserManager implementation (like Kenai Realm is),
the logs would be polluted with a lot of exceptions
on every log-in and UI use (status resource).

A Realm without UserManager is still serving the purpose
well (authc and authz would work just fine), but
operations like "user lookup" all fail due to lack
of UserManager.