NXCM-5406: Make nx-staging-plugin mvn3.1 friendly
It turned out that the plugin itself is okay
with mvn3.1, but it's the nexus-core-client that
used old Aether for some version comparisions.

Luckily, these ops are well hidden behind the Condition
iface, so for now copy paste of relevant
Aether classes happened into the client
(and removal of Aether dependency).

This makes nexus-staging-maven-plugin happy in
mvn 3.1