NEXUS-5772: Added UT
And it was a painful experience, as MimeUtil loves
static fields and states, and making it
"reboot" is PITA.

Also, initial implementation was actually wrong
(discovered while doing UT), as the Opendesktop
mime detector implicitly uses static/singleton
MimeUtil2 to add loaded mime entries, unless
is explicitly registered.... (and the static
one is still filled up too, so duplication
happens).

Hardly to be fixed, unless forked.