NXCM-5433: Attributes should never be thrashed
Since attributes were moved into LS (since 2.0),
on deletion of a file item, it's attributes were always
thrashed (content fate depends on actual DeleteOperation
in effect).

This was a long outstanding bug, as we do support content
restoration from thrash, but not attribute restoration.

After this change:
* attributes of deleted item(s) are lost forever (never thrashed)
* content of deleted item(s) are either thrashed or shredded
(depends on DeleteOperation in effect)