NEXUS-5836: Browse Index Tree fix
When pulling out the MI TreeView from direct
use in REST API (as DTO), during conversion
the resulting payload was slightly different
and that made UI think it deals with leafs, and
never asked for siblings.

Problem actually lies in UI, that neglects the
leaf boolean flag, but it actually tested
for children field (was null, after change
it was empty list, now is again null). Ultimately
it's the leaf flag that should be used to test is a
node a leaf or not, not the value of children array
(UI assumes if children == null, is not leaf,
otherwise it is).

The REST resource did work, and even IT did pass
as the IT uses the node:leaf and node:chilren
properties as should