On fresh cached files no UID lock was applied.
As revealed by debugging: with empty cache, a content
request that made Nx go remote, fetch, cache and serve
the content, the StorageFileItemRepresentation got a
file item which ContentLocator was not wrapped with
ReadLockingContentLocator (but it was plain FileContentLocator).

This means, that files served up like these (or processed internally)
were prone to deletion by some other thread, like request
for deletion of a parent directory or such.