Added plx XML to have component from here override MI component.
As before NexusArtifactPackagingMapper was PLX component,
it was overriding the ArtifactPackagingMapper from MI
just fine. But as everything got "sisuized", it
stopped happening, making the Nx mapper redirection
defunct (and test Nexus5525IndexerManagerTest did
signal that).

Fixing by adding a simple PLX XML to have it included in
final JAR, and to have the Plexus component override
happen again.