NEXUS-6109: Added new method for better precision.
Typically, JAR files are actually ZIP containers, and
Tika will recognize it as ZIP, but we did "reroute" MimeUtil2
and was showing application/java-archive in those cases.

So, a new method helps Tika guess better, as it is able to
use not only content, but also the file name to fine tune
the guesswork.