NEXUS-6099: Nexus1563ExternalRealmsLoginIT configuration tweak
In old Nx config, the security information was embedded into
nexus.xml, later was pulled out into separate file. Not much
ITs depended on that, but this one was. Simply put, the information
that Simple Realm is needed for this IT was lost during "bare"
upgrade of nexus.xml from ancient version, as security was not
upgraded in this case (is happening in full Nx instance via
eventing).