NEXUS-6099: Making Kenai Realm config lazy again
As it's trivially simple: is read only, and it interferes
with Kenai ITs that boot nexus to be able to create kenai-role
that is already in eagerly loaded config (causing problems
and Kenai plugin realm being discarded).

This way, Kenai config is lazy again, that does not stirs
any water in production, but makes ITs work again.