Do not swallow runtime exceptions
This is fix for the original commit
618bc6d2f262b37d40e3fbc3491952a6455834a6

The intent is to not allow AbstractProxyRepository
to swallow runtime exceptions and not have any
trace of them. The previous commit was done
as at one moment my code changes introduced
NPE in RRS, but it went completely unnoticed
and was revealed only by debugging the code
as nothing in logs showed by remote
retrieval failed.

This change should make it visible
from logs, when we deal with cases like
that, and should allow us early detection
of problems like those above, but also
should make problem detection on live instances
easier, without the need to run Nx in
DEBUG mode.