Make runtime ex logging happen at warn level.
Still, the point of logging here was to not swallow
and allow for these to happen traceless, hence
we do log them.

Still, these do not (and were not before change either)
breaking the execution chain here, as expected.

This is merely to have some insight, instead
of wonder why a proxy refuses to remotely
fetch something and it hide actual reasons
(like it was an NPE in my case)