NEXUS-6197: LVO should upgrade model version 1.0.0
Even if it is "old" and should be de-supported by
work done in NEXUS-6099, it seems LVO model version
was not really managed by Nx, and we still have
instances that are 2.x but LVO XML model version
is left at 1.0.0.

This change will just treat 1.0.0 as 1.0.1 as
those two are compatible (latter got new fields
with default values).

Added UT to exercise upgrades from 1.0.0 and
1.0.1.