Fixing ugly NPE/log pollition
If someone (by browser or curl) accesses the ErrorPageServlet.
As this request would be "unexpected", an NPE happened in it,
that caused nasty stack trace in log, and then ErrorFilter
would kick in, spitting the stack trace again (double NPE
in logs).

It was simply matter of Integer/int autoboxing and casting.
Now a simple check is added, to detect when page is
accessed externally and providing somewhat sensible
404 not found response instead of 500 Server Error
(and stack trace in log).