NEXUS-6249: Added UT verifying change
The UT simply forces Nexus to re-fetch the same file
(content or layout does not matter, it only matters it
happens in a Maven repository) and asserts that as
many times we change "version" of the content, the
retrieved digest will change too.

Before fixing this issue, Nexus would stick to the
remote hash retrieved the very first time.