NEXUS-6294: Deleted file handles.
While not quite same as NEXUS-6172, a bit different
approach was made here. Writer is kept open during the
lifetime of timeline component, and readers are now
opened per-request. This should be "light" operation
as it's done with applyDeletes=false.

Before, the searcherManager opened a reader at
component creation (nexus boot), and that
reader was kept as is even if nobody touched
feeds (reader keeps a "snapshot" of index frozen
at the point in time where it was open).