NEXUS-6591: remove old plugin ITs
I tried to salvage them, but the use cases and the
REST API -- while similar as they spun off same model --
diverged in way that the ITs are not fit to Pro LDAP

Also, Pro ITs basically test the Pro LDAP plugin
while these were testing the OSS, now removed plugin.

In behaviour, again, are similar, but there are differences
in edge cases (ie. no config).

Not worth the time