Fixing the build for two pass compilation.
Due to GROOVY-4287, two pass compilation of groovy class
with nested static classes are not possible, unless
java+groovy (all the code) is compiled in one pass.

For now, this bug makes eclipse to report the class
as broken (those importing Type), but any other attempt
for incremental compilation would fail too.

Simplest fix is to "rewrite" the SupportBundle to
plain Java.