NEXUS-6838: "Rude" fix, adding SNI to TLS/HTTPS
This is the "rude" way as described in HTTPCLIENT-1119.

Other solution is described in other comment, in short: create
"plain" socket, and then wrap that socket with SSLSocket, but
for that to work you'd need to have host dedicated SSLContext
(for each HOST you access).

SNI works only on Java7+