Make DefaultFSPeer use new DirSupport method with proper filter
As in case of move/delete, it would not only delete the
repo content, but also repo "metadolders" like ".meta"
(less of a problem), and ".nexus" where trash is, basically
performing a "move into a subdir" of directory. This ended
up in endless cycle, broken when OS refused file creation
for too long pathnames.