DirSupport move simplification
The "fallback" was needed only in one case, when
a dir was moved to it's subdir. But, to achieve
that one must declare excludeFilter. Hence,
from now on, WITHOUT filter that is impossible,
and in the same time, other IOEx from Files.move
should not be swallowed (ie. move over non-empty dir)