NEXUS-7671: Remove irrelevant UTs
That were testing feature purge-by-type/by-subtype and it's combinations
that were never actually used in production. Timeline interface does not
support this, only the indexer and it was tested only in the indexer
UTs, nowhere else.

It's broken as deletion happens per-term, and term includes timestamp only
but not the type nor subtype.