Making legacy REST perform less queries against QZ
But this will need to be reworked. The problem is caused by
transient (one-off, aka bg-tasks) that dissapear when they are
done and geters on TaskInfo start throwing ISEx-es.

This will need to be reworked: either TaskInfo of those tasks
needs to cache sorta (as config and other thing for one-off
tasks cannot change anyway), or not remove volatile tasks (have
some sort of "gc" doing that), or...