Handle properly premature cancelation
At the moment QZ "scheduler thread" did not yet pick up
the task to be run, and is scheduled to be run at this moment
(either by "now" or it triggers now), there is a window
when the task might be canceled (prematurely). This should
not throw UnableToInterruptJobException as cancelation in
this case can be handed in NexusTaskJobSupport just fine,
even if the NX task is not cancelable.