npm contentUri should always end in slash
In order for npm cli version 2.1.7 or higher to be happy, registry url passed to it must end in slash. For consistency, I propose to change the contentUri method to enforce this. This helps the PublishIT we have to pass the correct registry url to the npm cli.

See the mentioned npm issue I filed to see if this was a regression in the npm cli.