NEXUS-8060: Fix for original change to expose attributes
The introduced switch made endless loops in storage, as same attribute is used
in both: attribute maintenance (to break loops) and for remote
reachability. Solved by moving the "switches" to IsRemotelyAccessibleAttribute
and "denormalizing" the attribute (as originally used IsMetacontentAttribute
is used in loop detection).

UT updated with changes, and also renamed as it also had "Abstract"
prefix, and hence, was not run as part of UTs in this module.

Ported from
f5e41fc29ad1b6f3768f65d48139863623373cb8