[NEXUS-7860] attach userId during NexusBaseHttpAuthenticationFilter.onLoginSuccess
This handles the case where a subject is not logged in at the time of SecurityFilter.executeChain,
but is logged in after a challenge is sent. As SecurityFilter is not re-executed we need to attach
the userId during NexusBaseHttpAuthenticationFilter.onLoginSuccess. This lines up with the two uses
of UserIdMdcHelper.set: once in SecurityFilter.executeChain and once in NexusWebSecurityManager.login
(which is called from NexusBaseHttpAuthenticationFilter for login attempts).