Forcibly evict bundles when stopping/restarting.
As part of the shutdown process we stop our extender/tracker when the context is destroyed,
which triggers removal events for any bundles it was tracking. Before this change we used
the bundle state to decide whether to evict the bundle (ie. remove its components from the
bean locator) and we ignored the eviction of any bundles during shutdown because of legacy
components which needed stability.

Now we forcibly evict bundles when we know we're stopping regardless of their actual state.
Note: some bundles can still be in active state at the time the tracker is stopped because
at the time the context is destroyed the framework start level has not yet dropped to zero.