Install StateGuardModule as early as possible in the startup sequence.
Now that nexus-core depends on nexus-repository (and therefore nexus-repository will be processed before nexus-core) it makes more sense to install the StateGuardModule in the root context module instead of waiting until nexus-core is activated, potentially missing out on enhancing prior components. This revealed an incorrect assumption in StorageFacetImpl that hooks would be registered when the facet was in NEW state. Since these hooks are registered by other facets after StorageFacetImpl has been initialised it should instead be checking for INITIALISED state (and use the state constants from FacetSupport).